feat(users): add password recover and mail confirmation on register
